improve gif dumping
Animation gets compressed into a godot res file representing an Animated Texture for faster an easier loading.  The res file contains all sprite data embedded, but is not readable by anything other than godot.  Its file is guaranteed to be larger than the original format since it is a compilation of PNG frames and timing metadata, which is less compressed than the original gif or webp's encoding.

The folder of frame data is also deleted after processing is complete, leaving just the source gif and new res file behind.