Directory confirmation/creation tweaks
Only the /emotes/ folder was being confirmed to exist, and it did so in privmsg rather than one of the emote classes, etc. The most logical place for it seems to be utils.gd, which everything else is using for loading and saving images, and has the side effect of also solving the issue that the profile images weren't being saved because the directory wasn't created.