Updating pronouns.gd to V1 API (#3)
* Updated pronouns.gd to use the V1 API, and implemented requested changes for cleanup of pronoun output

* Directory confirmation/creation tweaks

Only the /emotes/ folder was being confirmed to exist, and it did so in privmsg rather than one of the emote classes, etc. The most logical place for it seems to be utils.gd, which everything else is using for loading and saving images, and has the side effect of also solving the issue that the profile images weren't being saved because the directory wasn't created.