Removed magick_dumps plugin in project settings
While the plugin is useful and I would recommend users get it as is mentioned in the readme, the base project doesn't come with it pre-installed which causes conflicts in the project settings (Trying to enable a plugin that doesn't exist). This should just remove it, and add the client_secret field that should go with the client_id field. I think both are needed? I may be wrong there.